# Improve ingestion process performance