--- a/entity/src/qualified_package.rs
+++ b/entity/src/qualified_package.rs
@@ -1,5 +1,6 @@
 use sea_orm::entity::prelude::*;
-use sea_orm::FromQueryResult;
+use sea_orm::{FromJsonQueryResult, FromQueryResult};
+use std::collections::HashMap;
 
 #[derive(Clone, Debug, PartialEq, Eq, DeriveEntityModel)]
 #[sea_orm(table_name = "qualified_package")]
@@ -7,7 +8,13 @@ pub struct Model {
     #[sea_orm(primary_key)]
     pub id: i32,
     pub package_version_id: i32,
+    pub qualifiers: Option<Qualifiers>,
 }
+
+#[derive(
+    Clone, Debug, PartialEq, Eq, serde::Serialize, serde::Deserialize, FromJsonQueryResult,
+)]
+pub struct Qualifiers(pub HashMap<String, String>);
 
 #[derive(Copy, Clone, Debug, EnumIter, DeriveRelation)]
 pub enum Relation {
@@ -16,19 +23,11 @@ pub enum Relation {
         from = "super::qualified_package::Column::PackageVersionId"
     to = "super::package_version::Column::Id")]
     PackageVersion,
-    #[sea_orm(has_many = "super::package_qualifier::Entity")]
-    Qualifiers,
 }
 
 impl Related<super::package_version::Entity> for Entity {
     fn to() -> RelationDef {
         Relation::PackageVersion.def()
-    }
-}
-
-impl Related<super::package_qualifier::Entity> for Entity {
-    fn to() -> RelationDef {
-        Relation::Qualifiers.def()
     }
 }