Down to:

```
[2024-04-17T13:09:09Z INFO  trustify_module_graph::graph::sbom::spdx::tests::perf] parse: 2s 530ms 515us 693ns
[2024-04-17T13:09:09Z INFO  trustify_module_graph::graph::sbom::spdx::tests::perf] ingest 1: 2ms 271us 191ns
[2024-04-17T13:09:09Z INFO  trustify_module_graph::graph::sbom::spdx::tests::perf] ingest 2: 13s 942ms 683us 307ns
[2024-04-17T13:09:09Z INFO  trustify_module_graph::graph::sbom::spdx::tests::perf] commit: 341us 276ns
[2024-04-17T13:09:09Z INFO  trustify_module_graph::graph::sbom::spdx::tests::perf] query: 4ms 501us 323ns
```

The two most impactful changes have been:

* using indexes for core lookup (yes, we pay for that with additional storage)
* flattening the "qualifier" map into a JSONB value, instead of having an additional table

Regarding the last change, I am not sure I fixed up all searches, but I hope the CI will help finding this out.

Without that change (aka "before"):

```
[2024-04-17T13:16:38Z INFO  trustify_module_graph::graph::sbom::spdx::tests::perf] parse 2s 427ms 764us 131ns
[2024-04-17T13:16:38Z INFO  trustify_module_graph::graph::sbom::spdx::tests::perf] ingest 46s 336ms 260us 291ns
[2024-04-17T13:16:38Z INFO  trustify_module_graph::graph::sbom::spdx::tests::perf] commit 355us 36ns
[2024-04-17T13:16:38Z INFO  trustify_module_graph::graph::sbom::spdx::tests::perf] query 4ms 477us 762ns
```