@ctron 2024-04-17
Did some more tests:

OpenShift 4.13: 

```
[2024-04-17T14:28:47Z INFO  trustify_module_graph::graph::sbom::spdx::tests::perf] parse: 40s 788ms 427us 523ns
[2024-04-17T14:28:47Z INFO  trustify_module_graph::graph::sbom::spdx::tests::perf] ingest 1: 44ms 874us 837ns
[2024-04-17T14:28:47Z INFO  trustify_module_graph::graph::sbom::spdx::tests::perf] ingest 2: 5m 19s 866ms 481us 581ns
[2024-04-17T14:28:47Z INFO  trustify_module_graph::graph::sbom::spdx::tests::perf] commit: 62ms 2us 9ns
[2024-04-17T14:28:47Z INFO  trustify_module_graph::graph::sbom::spdx::tests::perf] query: 4ms 90us 838ns
```

RHEL 9.4.0

```
[2024-04-17T14:41:30Z INFO  trustify_module_graph::graph::sbom::spdx::tests::perf] parse: 18s 357ms 70us 328ns
[2024-04-17T14:41:30Z INFO  trustify_module_graph::graph::sbom::spdx::tests::perf] ingest 1: 7ms 120us 916ns
[2024-04-17T14:41:30Z INFO  trustify_module_graph::graph::sbom::spdx::tests::perf] ingest 2: 1m 54s 954ms 576us 696ns
[2024-04-17T14:41:30Z INFO  trustify_module_graph::graph::sbom::spdx::tests::perf] commit: 28ms 993us 731ns
[2024-04-17T14:41:30Z INFO  trustify_module_graph::graph::sbom::spdx::tests::perf] query: 4ms 563us 26ns
```

IIRC it was the RHEL 9.4.0 one that took ~57mins when it tried it. I doubt the improvement is that good. So either something's wrong, or there are additional factors (like more SBOMs) which slow down the ingestion.